+++ includes/plugins/advanced-custom-fields/advanced-custom-fields.php
@@ -10,7 +10,7 @@
  * @copyright  WebMan Design, Oliver Juhas
  *
  * @since    2.0
- * @version  2.0
+ * @version  2.5.4
  *
  * Contents:
  *
@@ -26,7 +26,7 @@
  * 1) Requirements check
  */
 
-	if ( ! function_exists( 'register_field_group' ) || ! is_admin() ) {
+	if ( ! function_exists( 'acf_add_local_field_group' ) || ! is_admin() ) {
 		return;
 	}
 
@@ -44,7 +44,7 @@
 	 * @link  https://www.advancedcustomfields.com/resources/register-fields-via-php/
 	 *
 	 * @since    2.0
-	 * @version  2.0
+	 * @version  2.5.4
 	 */
 	function wm_acf_register_field_group() {
 
@@ -52,12 +52,12 @@
 
 			// Register metabox: Menu page template
 
-				register_field_group( array(
+				acf_add_local_field_group( array(
 					'id'     => 'wm_page_template_menu',
 					'title'  => esc_html__( 'Food menu page template options', 'auberge' ),
-					'fields' => array (
+					'fields' => array(
 
-						array (
+						array(
 							'key'             => 'wm_food_menu_section',
 							'label'           => esc_html__( 'Food menu section', 'auberge' ),
 							'name'            => 'food_menu_section',
@@ -73,17 +73,17 @@
 						),
 
 					),
-					'location' => array (
+					'location' => array(
 
-						array (
+						array(
 
-							array (
+							array(
 								'param'    => 'post_type',
 								'operator' => '==',
 								'value'    => 'page',
 							),
 
-							array (
+							array(
 								'param'    => 'page_template',
 								'operator' => '==',
 								'value'    => 'page-template/_menu.php',
@@ -92,24 +92,21 @@
 						),
 
 					),
-					'options' => array (
-						'position'       => 'normal',
-						'layout'         => 'default',
-						'hide_on_screen' => array(),
-					),
-					'menu_order' => 0,
+					'menu_order'     => 0,
+					'position'       => ( function_exists( 'has_blocks' ) ) ? ( 'side' ) : ( 'normal' ),
+					'style'          => 'default',
 				) );
 
 
 
 			// Register metabox: Quote post format
 
-				register_field_group( array(
+				acf_add_local_field_group( array(
 					'id'     => 'wm_post_format_quote',
 					'title'  => esc_html__( 'Quote post format options', 'auberge' ),
-					'fields' => array (
+					'fields' => array(
 
-						array (
+						array(
 							'key'           => 'wm_quote_source',
 							'label'         => esc_html__( 'Quote source', 'auberge' ),
 							'name'          => 'quote_source',
@@ -124,17 +121,17 @@
 						),
 
 					),
-					'location' => array (
+					'location' => array(
 
-						array (
+						array(
 
-							array (
+							array(
 								'param'    => 'post_type',
 								'operator' => '==',
 								'value'    => 'post',
 							),
 
-							array (
+							array(
 								'param'    => 'post_format',
 								'operator' => '==',
 								'value'    => 'quote',
@@ -143,24 +140,21 @@
 						),
 
 					),
-					'options' => array (
-						'position'       => 'normal',
-						'layout'         => 'default',
-						'hide_on_screen' => array(),
-					),
 					'menu_order' => 0,
+					'position'   => ( function_exists( 'has_blocks' ) ) ? ( 'side' ) : ( 'normal' ),
+					'style'      => 'default',
 				) );
 
 
 
 			// Register metabox: Front page
 
-				register_field_group( array(
+				acf_add_local_field_group( array(
 					'id'     => 'wm_page_front',
 					'title'  => esc_html__( 'Front page options', 'auberge' ),
-					'fields' => array (
+					'fields' => array(
 
-						array (
+						array(
 							'key'           => 'wm_banner_text',
 							'label'         => esc_html__( 'Banner text', 'auberge' ),
 							'name'          => 'banner_text',
@@ -175,17 +169,17 @@
 						),
 
 					),
-					'location' => array (
+					'location' => array(
 
-						array (
+						array(
 
-							array (
+							array(
 								'param'    => 'post_type',
 								'operator' => '==',
 								'value'    => 'page',
 							),
 
-							array (
+							array(
 								'param'    => 'page_type',
 								'operator' => '==',
 								'value'    => 'front_page',
@@ -194,24 +188,21 @@
 						),
 
 					),
-					'options' => array (
-						'position'       => 'normal',
-						'layout'         => 'default',
-						'hide_on_screen' => array(),
-					),
 					'menu_order' => 0,
+					'position'   => ( function_exists( 'has_blocks' ) ) ? ( 'side' ) : ( 'normal' ),
+					'style'      => 'default',
 				) );
 
 
 
 			// Register metabox: Featured post
 
-				register_field_group( array(
+				acf_add_local_field_group( array(
 					'id'     => 'wm_post_featured',
 					'title'  => esc_html__( 'Featured post options', 'auberge' ),
-					'fields' => array (
+					'fields' => array(
 
-						array (
+						array(
 							'key'           => 'wm_banner_text',
 							'label'         => esc_html__( 'Banner text', 'auberge' ),
 							'name'          => 'banner_text',
@@ -226,11 +217,11 @@
 						),
 
 					),
-					'location' => array (
+					'location' => array(
 
-						array (
+						array(
 
-							array (
+							array(
 								'param'    => 'post_type',
 								'operator' => '==',
 								'value'    => 'post',
@@ -239,14 +230,11 @@
 						),
 
 					),
-					'options' => array (
-						'position'       => 'normal',
-						'layout'         => 'default',
-						'hide_on_screen' => array(),
-					),
 					'menu_order' => 1,
+					'position'   => ( function_exists( 'has_blocks' ) ) ? ( 'side' ) : ( 'normal' ),
+					'style'      => 'default',
 				) );
 
 	} // /wm_acf_register_field_group
 
-	add_action( 'init', 'wm_acf_register_field_group' );
+	add_action( 'init', 'wm_acf_register_field_group', 100 );